src/main/java/com/blinkstore/controllers/errors/BuyItemsErrorHandler.java
package com.blinkstore.controllers.errors;

import org.springframework.http.HttpStatus;
import org.springframework.web.bind.MethodArgumentNotValidException;
import org.springframework.web.bind.annotation.ControllerAdvice;
import org.springframework.web.bind.annotation.ExceptionHandler;
import org.springframework.web.bind.annotation.ResponseBody;
import org.springframework.web.bind.annotation.ResponseStatus;

@ControllerAdvice
public class BuyItemsErrorHandler {

	@ExceptionHandler(MethodArgumentNotValidException.class)
	@ResponseStatus(HttpStatus.BAD_REQUEST)
	@ResponseBody
	public BuyItemErrorResponse handleBuyItemsValidationError() {
		return new BuyItemErrorResponse("We are sorry, but we are unable to realise your order. We do not have enough commodity on stock");
	}
}
